# app_backend/swagger.yaml
openapi: 3.0.3
info:
  title: API de Gestión de Inventario, Préstamos y Penalizaciones
  description: >
    Esta API maneja la autenticación de usuarios, administración de perfiles, 
    control centralizado de stock (materiales, herramientas, placas, libros), 
    flujo de préstamos/reservas mediante validación por QR, gestión de penalizaciones 
    y generación de reportes/estadísticas.
  version: 2.0.0
servers:
  - url: /
    description: Servidor Base Local
paths:
...
  /api/penalizaciones:
    get:
      tags:
        - Penalizaciones
      summary: Lista general de penalizaciones activas e históricas
      description: Admin y bibliotecario pueden listar todas las penalizaciones; alumno y profesor solo pueden consultar las propias con usuario_id.
      security:
        - BearerAuth: []
      parameters:
        - name: usuario_id
          in: query
          required: false
          description: Identificador del usuario para filtrar penalizaciones.
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Listado obtenido.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PenaltyDetail'
        '400':
          description: Parámetros de consulta inválidos.
        '403':
          description: Prohibido. El usuario autenticado no puede consultar penalizaciones ajenas.

    post:
      tags:
        - Penalizaciones
      summary: Genera una nueva penalización manual a un alumno
      description: Admin, bibliotecario o profesor pueden crear una penalización manual.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PenaltyCreate'
      responses:
        '201':
          description: Penalización creada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PenaltyDetail'
        '400':
          description: Error en datos.
        '403':
          description: Prohibido.
  /api/penalizaciones/{penalty_id}:
    parameters:
      - name: penalty_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags:
        - Penalizaciones
      summary: Obtiene el detalle de una penalización específica
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Detalle obtenido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PenaltyDetail'
        '404':
          description: Penalización no encontrada.

    put:
      tags:
        - Penalizaciones
      summary: Modifica o levanta una penalización reemplazando el registro completo
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PenaltyUpdate'
      responses:
        '200':
          description: Penalización reemplazada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PenaltyDetail'
        '404':
          description: Penalización no encontrada.

    patch:
      tags:
        - Penalizaciones
      summary: Actualiza parcialmente los datos de una penalización
      description: Ajusta severidad, agrega notas o la marca como resuelta.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PenaltyPatchUpdate'
      responses:
        '200':
          description: Penalización actualizada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PenaltyDetail'
        '404':
          description: Penalización no encontrada.

  # ==========================================
  # 7. Normativas (/api/normativas)
  # ==========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PenaltyDetail:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        loanId:
          type: integer
          format: int64
          nullable: true
        reason:
          type: string
        status:
          type: string
          enum: [Activa, Levantada]
        severity:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
    PenaltyCreate:
      type: object
      required:
        - usuario_id
        - reason
      properties:
        usuario_id:
          type: integer
          format: int64
        loanId:
          type: integer
          format: int64
        reason:
          type: string
    PenaltyUpdate:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [Activa, Levantada]
        resolutionNotes:
          type: string
    PenaltyPatchUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [Activa, Levantada]
        severidad:
          type: string
          example: Alta
        notes:
          type: string
...